% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{est_eb}
\alias{est_eb}
\title{Competitors: Empirical Bayes (EB) estimator}
\usage{
est_eb(beta_int, beta_ext, V_int, p)
}
\arguments{
\item{beta_int}{Internal OLS estimate}

\item{beta_ext}{External point estimate}

\item{V_int}{Internal covariance matrix of beta_int}

\item{p}{Number of parameters}
}
\value{
EB-integrated coefficient vector
}
\description{
Competitors: Empirical Bayes (EB) estimator
}
