% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{est_js_plus}
\alias{est_js_plus}
\title{Competitors: Positive-part James-Stein (JS+) estimator}
\usage{
est_js_plus(beta_int, beta_ext, X_int, sigma2_int, p)
}
\arguments{
\item{beta_int}{Internal OLS estimate}

\item{beta_ext}{External point estimate}

\item{X_int}{Internal design matrix}

\item{sigma2_int}{Internal error variance}

\item{p}{Number of parameters}
}
\value{
JS+-integrated coefficient vector
}
\description{
Competitors: Positive-part James-Stein (JS+) estimator
}
