% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{est_partial}
\alias{est_partial}
\title{Partial data-sharing estimator}
\usage{
est_partial(beta_int, beta_ext, eta)
}
\arguments{
\item{beta_int}{Internal point estimate (p-vector)}

\item{beta_ext}{External point estimate (p-vector)}

\item{eta}{Tuning parameter (scalar >= 0)}
}
\value{
Integrated coefficient vector (p-vector)
}
\description{
Computes the scalar-weighted average:
\eqn{beta_int / (1 + eta) + eta * beta_ext / (1 + eta)}
}
