% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.mainger}
\alias{predict.mainger}
\title{Predict method for mainger objects}
\usage{
\method{predict}{mainger}(object, newdata, ...)
}
\arguments{
\item{object}{A mainger object}

\item{newdata}{New design matrix (n_new x p)}

\item{...}{Additional arguments}
}
\value{
Predicted response vector (n_new x 1)
}
\description{
Predict method for mainger objects
}
