% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.mainger}
\alias{print.mainger}
\title{Print method for mainger objects}
\usage{
\method{print}{mainger}(x, ...)
}
\arguments{
\item{x}{A mainger object}

\item{...}{Additional arguments}
}
\description{
Print method for mainger objects
}
