% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.mainger}
\alias{summary.mainger}
\title{Summary method for mainger objects}
\usage{
\method{summary}{mainger}(object, n_top = 10, ...)
}
\arguments{
\item{object}{A mainger object}

\item{n_top}{Number of top coefficients to show (default 10)}

\item{...}{Additional arguments (not used)}
}
\description{
Summary method for mainger objects
}
